+++ content/themes/make/header.php
@@ -6,12 +6,6 @@
 <!--[if lte IE 9]><html class="no-js IE9 IE" <?php language_attributes(); ?>><![endif]-->
 <!--[if gt IE 9]><!--><html class="no-js" <?php language_attributes(); ?>><!--<![endif]-->
 	<head>
-		<meta charset="<?php bloginfo( 'charset' ); ?>">
-		<meta name="viewport" content="width=device-width, initial-scale=1">
-		<meta http-equiv="X-UA-Compatible" content="IE=edge" />
-
-		<title><?php wp_title( '|', true, 'right' ); ?></title>
-
 		<?php wp_head(); ?>
 	</head>